backend_rag/pyproject.toml
[project]
name = "backend-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "fastembed>=0.7.4",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "uvicorn>=0.40.0",
]
